requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
pydantic==1.8.2
python-dateutil==2.8.2
pyOpenSSL==20.0.1
certifi==2021.10.8
gunicorn==20.1.0
python-magic==0.4.24
APScheduler==3.8.1
loguru==0.5.3
pytz==2021.1
psutil==5.9.0
